requirements.txt
langchain
openai
faiss-cpu
pandas
numpy
tqdm
python-dotenv
